Fix tests
We need some dependency additions (and a .babelrc) to get the tests
passing since f03a0891945b56807446f7590.

Additionally, the data returned from the API now has "unknown" for some
average lifespan(s), which means that we have to explicitly check for
that, because "unknown" cannot be coerced to `GraphQLInt`.