Add "build-public" script to build an independent static app
Consists of:

- Simple index.html file.
- Main app bundle that loads dependencies and renders GraphiQL:
  - "fetch" and "react" load in parallel, then "react-dom", and finally
    "graphiql".
- "schema" bundle that exposes the app schema, "graphql" executor, and
  cached data; this is loaded separately from within the GraphiQL
  fetcher callback.

The main app bundle is about 56KB in size (15KB gzipped) and the
schema/data bundle is 1.0MB (152KB gzipped).

I'll have follow-up commits to this, including a script that actually
deploys to the gh-pages branch, but first I have to investigating
whether there'll be any conflict with our existing gh-pages for
graphql.github.io.