Add "deploy" command to package.json
This runs the static build then commits the result to the gh-pages
branch.

Once pushed, it will be live at:

  http://graphql.github.io/swapi-graphql

I checked how this works with our other organization repos. It works
fine for the main graphql.github.io repo to write the files into the
root of domain, and other repos will wind up in sub-directories; as long
as the directory names don't conflict, everything should be fine.