Fix some Flow errors
There are still some remaining errors that may be trickier to solve;
specifically, it is complaining like this:

    CommonJS exports of "./types/starship". This type is incompatible
    with the expected return type of ... GraphQLObjectType

The issue being that:

    import StarshipType from './starship';

works (ie. importing a `default` export from an ES6 module into another
ES5 module), but this:

    const StarshipType = require('./types/starship');

is rejected by Flow because it thinks that a `require` in this case
should return an object with the shape:

    {
      default: StarshipType,
    }

So, I can make the Flow errors go away by changing the `require` to:

    const StarshipType = require('./types/starship').default;

But that doesn't work at runtime because Babel is evidently not
transforming the `require` (or perhaps the module) in the expected way.
So, going to look into a Babel upgrade/tweak before going further with
this.