Fix remaining Flow errors
Let's get rid of babel-plugin-add-module-exports, which makes Babel 6
have the (looser) Babel 5 behavior. This means we have to add a few
`.default` suffixes in some places, but it allows us to keep Flow (which
wants to more strictly conform to the spec/convention) happy. For more
context on this, see the previous commit.

This required me to fix the exports of `schema-proxy.js`, which I
actually had wrong before but I was getting away with it thanks to the
plug-in.